chore: split error to client and server sides (#27672)
