+++ packages/playwright-core/src/client/worker.ts
@@ -23,7 +23,7 @@ import type { BrowserContext } from './browserContext';
 import type * as api from '../../types/types';
 import type * as structs from '../../types/structs';
 import { LongStandingScope } from '../utils';
-import { TargetClosedError } from '../common/errors';
+import { TargetClosedError } from './errors';
 
 export class Worker extends ChannelOwner<channels.WorkerChannel> implements api.Worker {
   _page: Page | undefined;  // Set for web workers.
