+++ packages/playwright-core/src/server/errors.ts
@@ -1,12 +1,11 @@
 /**
- * Copyright 2018 Google Inc. All rights reserved.
- * Modifications copyright (c) Microsoft Corporation.
+ * Copyright (c) Microsoft Corporation.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,23 +14,43 @@
  * limitations under the License.
  */
 
+import type { SerializedError } from '@protocol/channels';
+import { isError } from '../utils';
+import { parseSerializedValue, serializeValue } from '../protocol/serializers';
+
 class CustomError extends Error {
   constructor(message: string) {
     super(message);
     this.name = this.constructor.name;
-    Error.captureStackTrace(this, this.constructor);
   }
 }
 
 export class TimeoutError extends CustomError {}
 
-export class TargetClosedError extends Error {
+export class TargetClosedError extends CustomError {
   constructor(cause?: string, logs?: string) {
     super((cause || 'Target page, context or browser has been closed') + (logs || ''));
-    this.name = this.constructor.name;
   }
 }
 
 export function isTargetClosedError(error: Error) {
   return error instanceof TargetClosedError || error.name === 'TargetClosedError';
+}
+
+export function serializeError(e: any): SerializedError {
+  if (isError(e))
+    return { error: { message: e.message, stack: e.stack, name: e.name } };
+  return { value: serializeValue(e, value => ({ fallThrough: value })) };
+}
+
+export function parseError(error: SerializedError): Error {
+  if (!error.error) {
+    if (error.value === undefined)
+      throw new Error('Serialized error must have either an error or a value');
+    return parseSerializedValue(error.value, undefined);
+  }
+  const e = new Error(error.error.message);
+  e.stack = error.error.stack || '';
+  e.name = error.error.name;
+  return e;
 }
