+++ packages/playwright-core/src/server/webkit/wkPage.ts
@@ -45,7 +45,7 @@ import { WKWorkers } from './wkWorkers';
 import { debugLogger } from '../../common/debugLogger';
 import { ManualPromise } from '../../utils/manualPromise';
 import { BrowserContext } from '../browserContext';
-import { TargetClosedError } from '../../common/errors';
+import { TargetClosedError } from '../errors';
 
 const UTILITY_WORLD_NAME = '__playwright_utility_world__';
